src/test/kotlin/com/example/graphqldgsdemo/datafetchers/CocktailsDataFetcherTest.kt
package com.example.graphqldgsdemo.datafetchers

import com.netflix.graphql.dgs.DgsQueryExecutor
import com.netflix.graphql.dgs.autoconfig.DgsAutoConfiguration
import org.assertj.core.api.Assertions.assertThat
import org.junit.jupiter.api.Test
import org.springframework.beans.factory.annotation.Autowired
import org.springframework.boot.test.context.SpringBootTest

@SpringBootTest(classes = [DgsAutoConfiguration::class, CocktailsDataFetcher::class])
class CocktailsDataFetcherTest {

    @Autowired
    lateinit var dgsQueryExecutor: DgsQueryExecutor

    @Test
    fun cocktails() {
        val titles: List<String> = dgsQueryExecutor.executeAndExtractJsonPath(
            """
            {
                cocktails {
                    name
                }
            }
        """.trimIndent(), "data.cocktails[*].name"
        )

        assertThat(titles).contains("Pina colada")
    }
}
